Fix the multiple implementations of the getQualifiedName() and getMethodSignature() that were retuning an odd string representation for method signatures. All implementations call a base getQualifiedName() implementation.
The TypeMirror.toString() method returns the fully qualified name of the type.
If the type is a method signature, it places the parameters list (parenteses)
before the return type (that is void if none), which is an odd convention for Java Code.
This way, we invert that order for a conventional representation of a method signature.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>