Adds a Parser.getAnnotationArgumentType() method that correctly parses the data type of an annotation argument.
When extracting the type information for a given annotation argument,
it was being returned the signature of the method in the interface that defines the annotation.
This method is used to internally set a value for the argument.
But we are interested only in knowing the data type of the argument.
This way, we need to get the return type of this method for that.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>