Fix #6
The new Doclet API has a different way to process each option.
Each Option instance is usually created as an anonymous class to implement
the process method. This method is called when a given command line option
matches the option name. The method is therefore accountable
to validate the option arguments.

Since creating an anonymous class is not usual nowadays,
the CustomOption class previously was changed to accept a
`BiPredicate<String, List<String>> argumentsProcessor`
function that enables each option instance to define
how the process method will be implemented,
without the need to create an anonymous class.

However, this argumentsProcessor was not being used yet.
Now, when an Option is created, the constructor receives
a BiPredicate that defines the implementation of the processor method.

There is no need for the Apache Commons CLI anymore,
(mainly because the doclet options are very simple)
and a lot of code could be removed, simplifying the implementation.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>